Split model configuration to separate classes
